SNOW-68721 Set the latest golang version to 1.12.x  - Change test configuration  - Remove test files for old golang versions.
