+++ .travis.yml
@@ -2,9 +2,11 @@ sudo: false
 language: go
 matrix:
   include:
-  - go: '1.10'
   - go: 1.11.x
   - go: 1.11.x
+    env: SNOWFLAKE_AZURE=true
+  - go: 1.12.x
+  - go: 1.12.x
     env: SNOWFLAKE_AZURE=true
 install:
 - |
