+++ priv_key_test_coding_helper.go
